Add support for spaces in BioNLP annotation and relation types
BioNLP format does not support spaces in annotation or relation type
names. This commit adds handling for such cases. When an annotation or
relation type has a space, it is replaced by a non-space character when
writing BioNLP formatted files (by default, the ‘^’ is used). When
reading BioNLP formatted files, the space placeholder (‘^’) is replaced
with a space.