Revised CoNLLCoref reader/writer to handle annotations with discontinuous spans
The CoNLLCoref writer now included a character with the chain
identifier to indicate pieces of a discontinuous annotation, e.g. (2a)

Also added the code that was used to update the coreference annotations
for CRAFT v3.1. This code removes identity chains that are of length 1,
applies manually curated changes to many of the chains, and merges
chains that share a mention. See the README.md file in the
CRAFT.git/coreference-annotation/ directory for details.

The ability to output APPOS relations in the CoNLLCoref format was
removed as the APPOS relations are not really chains, and therefore the
format is not suitable.