Further discontinuous span validation
Mapping spans to token boundaries can also cause instances of nested
discontinuous spans, so the validation code for discontinuous spans was
added to the CoNLL-Coref document writer. There was a case in
16628246.xml (coreference annotations) where "7.5 dbc embryos" was
annotated as "7" .. "5 dbc embryos". In this case the "7" maps to the
"7.5" token and the "5" also maps to the "7.5" token, so the final
annotation had two instances of the "7.5" token span. Seems like the
original annotation was faulty, i.e. the "7" .. "5" split, but it
happened, so a fix was required.